fixed bash script

+++ run.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
-rm -r build
+if [ -d "build" ]; then
+	rm -r build
+fi
 mkdir build
 cd build
 cp ../makefile .
 make
 ./ConnectiCab
-
